Move basic visitor to it's own directory
